requirements.txt
fastapi
uvicorn
requests
pyjwt
python-multipart
